refactor: add re-export and change import-order for all code
