+++ src/components/Courses/components/index.ts
@@ -1,0 +1,2 @@
+export * from './CourseCard';
+export * from './SearchBar';
