+++ src/components/CreateCourse/components/index.ts
@@ -1,0 +1,4 @@
+export * from './AuthorForm';
+export * from './AuthorItem';
+export * from './AuthorList';
+export * from './CreateCourseForm';
