+++ src/components/Header/components/HeaderWrapper/HeaderWrapper.tsx
@@ -1,4 +1,5 @@
 import { ReactNode } from 'react';
+
 import styles from './styles.module.scss';
 
 interface HeaderWrapperProps {
