+++ src/components/Header/components/index.ts
@@ -1,0 +1,3 @@
+export * from './HeaderWrapper';
+export * from './Logo';
+export * from './Username';
