+++ src/helpers/index.ts
@@ -1,3 +1,4 @@
 export { formatFullName } from './formatFullName';
 export { formatCreatedDate } from './formatCreatedDate';
 export { getDurationTime } from './getDurationTime';
+export { getAuthorIdsFromCourses } from './getAuthorIdsFromCourses';
